Move PricingCard and FeatureCard from index.tsx to new component files
